feat: download exe from gh release
