+++ keadm/cmd/keadm/app/cmd/util/nssm_test.go
@@ -1,3 +1,5 @@
+//go:build windows
+
 package util
 
 import (
